# Conflicts:
#	DrinkMaster/DrinkMaster/Controllers/GameStateModelsController.cs
#	DrinkMaster/DrinkMaster/Controllers/PlayerModelsController.cs
#	DrinkMaster/DrinkMaster/Views/Shared/_Layout.cshtml

+++ DrinkMaster/DrinkMaster/Models/PlayerDrinkModel.cs
@@ -12,7 +12,7 @@ namespace DrinkMaster.Models
         [Required]
         public string Name { get; set; }
         [Required]
-        public int AlcoholPercentage { get; set; }
+        public double AlcoholPercentage { get; set; }
         [Required]
         public int DrinkQuantity { get; set; }
         [Required]
